requirements.txt
streamlit
pytesseract
pillow
requests
kraken